Make runSearch injectable into MLSearchContainer
This is a best practice, since our MLSearchContainer is no longer
dependent on a specific module. I like the idea of keeping such
dependencies isolated in a very high-level container, like App.

The tests provided the initial motivation for this (so we can provide a
mock instead of a real search implementation). But I think this does
point to a higher level of clean code.