Extract runSearch as actionCreator
With a test, though I am not totally happy with this test setup.

I increasingly think that we should treat actionCreators and selectors
as the public API for the module, and the reducers as an implementation
detail.

This may argue for organizing modules as "ducks":
https://github.com/erikras/ducks-modular-redux

It may also argue for testing actionCreators and selectors together, and
leaving reducers as a black box. They could be unit-tested in case we
want to cover complexity, but this should be a secondary concern, since
such tests will be tied to *implementation* rather than *behavior*, and
thus brittle when we want to refactor. See discussion here:
https://github.com/reactjs/redux/issues/2179

Such an approach would hopefully allow us to get rid of the mockStore
dependency in the middle - we could use a real store.

Also, should the specific search library to be used be passed in as an
argument to runSearch? This would let us get rid of `nock`, since we
would not have to actually make any http calls at all, and instead
provide a test implementation of the search api adapter.