Move rendering details down to <App>
The pattern that I used in agile-mastering-toolkit, and which I think
makes sense, is to only put highest-order components (like <Provider>
and <BrowserRouter>) into index.js. For everything specific to rendering
normal components within that, push it down into <App>.

This avoids repeating things like wrapping selectors / actions. That
should be handled within App or lower in the component tree.