Prevent create-react-app from clearing console
At the cost of not being able to detect when there is a port collision
on the port being used for the UI (3000 by default).

It would be nice if we could turn off create-react-app's console
clearing behavior while maintaining the ability to interact with the
console. See
https://github.com/facebook/create-react-app/issues/2495#issuecomment-344537005
and comments below it