GROVE-429: pass 'id' to DetailContainer instead of 'uri'
BREAKING CHANGE! The upgrade instruction is to replace these lines, plus
update grove-core-react-components and
grove-core-react-redux-containers, once they are released.

Unfortunately, we can't rely on the React Router to be consistent when
it automatically decodes params. Another option would be to use a query
string instead, but this seems to work fine.